+++ tests/spec/libsass_closed_issues/mod.rs
@@ -1493,7 +1493,6 @@ fn issue_1162() {
 
 // From "sass-spec/spec/libsass-closed-issues/issue_1163.hrx"
 #[test]
-#[ignore] // wrong result
 fn issue_1163() {
     assert_eq!(
         rsass(